Equals, StrictEquals operations with basic unit tests
