+++ src/commonMain/kotlin/operations/logic/Equals.kt
@@ -1,0 +1,13 @@
+package operations.logic
+
+import LogicOperation
+import asList
+import operations.numeric.compare.ComparingOperation
+
+internal object Equals : LogicOperation, ComparingOperation {
+    override val key: String = "=="
+
+    override fun invoke(expression: Any?, data: Any?): Boolean = with(expression.asList) {
+        compareListOfTwo(this) { first, second -> first == second }
+    }
+}
