+++ src/commonMain/kotlin/operations/logic/NotStrictEquals.kt
@@ -1,0 +1,17 @@
+package operations.logic
+
+import LogicOperation
+import asList
+import comparableList
+import operations.numeric.compare.ComparingOperation
+import operations.numeric.compare.StrictComparingOperation
+import secondOrNull
+
+internal object NotStrictEquals : LogicOperation, StrictComparingOperation {
+    override val key: String = "!=="
+
+    override fun invoke(expression: Any?, data: Any?): Boolean = with(expression.asList) {
+        compareListOfTwoOrDefault(values = this, default = true) { first, second -> first != second }
+    }
+}
+
