+++ src/commonMain/kotlin/operations/numeric/compare/GreaterThan.kt
@@ -1,4 +1,4 @@
-package operations.numeric
+package operations.numeric.compare
 
 import LogicOperation
 import asList
