+++ src/commonMain/kotlin/operations/numeric/compare/LessThan.kt
@@ -1,9 +1,9 @@
-package operations.numeric
+package operations.numeric.compare
 
 import LogicOperation
 import asList
 
-internal object LessThan : LogicOperation, ComparingOperation {
+internal object LessThan : LogicOperation, RangeComparingOperation {
     override val key: String = "<"
 
     override fun invoke(expression: Any?, data: Any?): Any =
