+++ src/commonMain/kotlin/operations/numeric/compare/LessThanOrEqualTo.kt
@@ -1,9 +1,9 @@
-package operations.numeric
+package operations.numeric.compare
 
 import LogicOperation
 import asList
 
-internal object LessThanOrEqualTo : LogicOperation, ComparingOperation {
+internal object LessThanOrEqualTo : LogicOperation, RangeComparingOperation {
     override val key: String = "<="
 
     override fun invoke(expression: Any?, data: Any?): Any =
