+++ src/commonTest/kotlin/operations/string/SubstrTest.kt
@@ -7,7 +7,7 @@ import io.kotest.datatest.withData
 import io.kotest.matchers.shouldBe
 
 class SubstrTest : FunSpec({
-    context("JsonLogic evaluation with Cat operation") {
+    context("JsonLogic evaluation with Substr operation") {
         withData(
             nameFn = { "Should apply ${it.data} on ${it.expression} result in ${it.result}" },
             // given
